refactor: split cell collecting and fetching
