+++ ckb/src/main/java/org/nervos/ckb/transaction/CollectIterator.java
@@ -1,0 +1,6 @@
+package org.nervos.ckb.transaction;
+
+/** Copyright © 2019 Nervos Foundation. All rights reserved. */
+public interface CollectIterator<CellInput, BigInteger, String> {
+  boolean hasNext(CellInput input, BigInteger capacity, String lockHash);
+}
