+++ ckb/src/main/java/org/nervos/ckb/transaction/CollectResult.java
