+++ example/src/main/java/org/nervos/ckb/MultiKeySingleSigTxExample.java
@@ -119,8 +119,8 @@ public class MultiKeySingleSigTxExample {
   }
 
   private static String getBalance(String address) throws IOException {
-    CellCollector cellCollector = new CellCollector(api);
-    return cellCollector.getCapacityWithAddress(address).divide(UnitCKB).toString(10);
+    CellFetcher cellFetcher = new CellFetcher(api);
+    return cellFetcher.getCapacityWithAddress(address).divide(UnitCKB).toString(10);
   }
 
   private static String sendCapacity(
